Refactor views into reusable components
