+++ RunTail/RunTail/Views/ActivityTabView.swift
@@ -324,129 +324,7 @@ struct ActivityTabView: View {
         }
     }
     
-    // MARK: - 헬퍼 뷰
-    
-    // 달력 날짜 셀
-    struct CalendarDayCell: View {
-        let date: Date
-        let isToday: Bool
-        let isActive: Bool
-        let isSelected: Bool
-        
-        private let calendar = Calendar.current
-        
-        var body: some View {
-            ZStack {
-                // 배경
-                if isToday {
-                    Circle()
-                        .fill(Color.rtPrimary)
-                        .frame(width: 36, height: 36)
-                } else if isActive {
-                    Circle()
-                        .fill(Color.rtPrimary.opacity(0.2))
-                        .frame(width: 36, height: 36)
-                } else if isSelected {
-                    Circle()
-                        .stroke(Color.rtPrimary, lineWidth: 1)
-                        .frame(width: 36, height: 36)
-                }
-                
-                // 날짜 텍스트
-                Text("\(calendar.component(.day, from: date))")
-                    .font(.system(size: 14))
-                    .foregroundColor(
-                        isToday ? .white :
-                            (calendar.component(.weekday, from: date) == 1 ? .red : .primary)
-                    )
-            }
-            .frame(height: 36)
-        }
-    }
-    
-    // 러닝 히스토리 아이템
-    struct RunningHistoryItem: View {
-        let run: Run
-        let onTap: () -> Void
-        
-        var body: some View {
-            Button(action: onTap) {
-                HStack(spacing: 16) {
-                    // 날짜 표시
-                    VStack(spacing: 4) {
-                        Text(monthString(from: run.runAt))
-                            .font(.system(size: 12))
-                            .foregroundColor(.gray)
-                        
-                        Text("\(Calendar.current.component(.day, from: run.runAt))")
-                            .font(.system(size: 18, weight: .bold))
-                        
-                        Text(yearString(from: run.runAt))
-                            .font(.system(size: 12))
-                            .foregroundColor(.gray)
-                    }
-                    .frame(width: 50)
-                    
-                    // 구분선
-                    Rectangle()
-                        .fill(Color.gray.opacity(0.3))
-                        .frame(width: 1, height: 50)
-                    
-                    // 러닝 정보
-                    VStack(alignment: .leading, spacing: 4) {
-                        Text(getDayOfWeek(date: run.runAt))
-                            .font(.system(size: 16, weight: .medium))
-                        
-                        // 거리와 시간
-                        let distance = run.trail.count > 0 ? 150 * Double(run.trail.count) : 0
-                        Text("\(Formatters.formatDistance(distance)) · \(Formatters.formatDuration(run.duration))")
-                            .font(.system(size: 14))
-                            .foregroundColor(.gray)
-                    }
-                    
-                    Spacer()
-                    
-                    // 페이스
-                    VStack(alignment: .trailing, spacing: 4) {
-                        Text("페이스")
-                            .font(.system(size: 12))
-                            .foregroundColor(.gray)
-                        
-                        Text(run.paceStr)
-                            .font(.system(size: 16, weight: .medium))
-                            .foregroundColor(.rtPrimary)
-                    }
-                }
-                .padding()
-                .background(Color.white)
-                .cornerRadius(16)
-                .shadow(color: Color.black.opacity(0.05), radius: 4, x: 0, y: 2)
-            }
-            .buttonStyle(PlainButtonStyle())
-        }
-        
-        // 요일 가져오기
-        private func getDayOfWeek(date: Date) -> String {
-            let formatter = DateFormatter()
-            formatter.dateFormat = "EEEE" // 요일 전체 이름
-            formatter.locale = Locale(identifier: "ko_KR")
-            return formatter.string(from: date)
-        }
-        
-        // 월 문자열
-        private func monthString(from date: Date) -> String {
-            let formatter = DateFormatter()
-            formatter.dateFormat = "MM월"
-            return formatter.string(from: date)
-        }
-        
-        // 년도 문자열
-        private func yearString(from date: Date) -> String {
-            let formatter = DateFormatter()
-            formatter.dateFormat = "yyyy"
-            return formatter.string(from: date)
-        }
-    }
+
     
     // MARK: - 헬퍼 메서드
